Added randomness to notifications
